[HYDRA-682] Refactor spark plugins into spark directory
