+++ spark-plugins/src/main/java/co/cask/hydrator/plugin/spark/ReferenceStreamingSource.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package co.cask.hydrator.plugin.batch.spark;
+package co.cask.hydrator.plugin.spark;
 
 import co.cask.cdap.api.Transactional;
 import co.cask.cdap.api.TxRunnable;
